refactor :  카테고리 조회, 검색 조회 기능 훅으로 묶어서 재사용
